// devui/src/app/shared/models/dto/research-request-searched-rule-dto.ts
export class ResearchRequestSearchedRuleDto {
    rrId: number;
    rrRuleMappingId: number;
    ruleId: number;
    ruleCode: string;
    ruleName: string;
    icmsId: string;
    changeType: number;
    reviewComments: string;
    ruleStatus: string;
    rrRuleStatus: number;
    statusId: number;
    ruleResponseIndicator: string;
    approvalStatus: string;
    userId:number;
    idIndicator: number;
    eclStageId: number;
    assignedTo: number;
    createdBy: number;
    poSubmitAction: string;
    cvpCpeRuleLink: boolean;
    eclRuleEngineId: number;
    engine: string;
    previousVersionLibraryRulePresent: boolean;
    
}